+++ lib/fast_file_picker.dart
@@ -1,0 +1,2 @@
+export 'src/fast_file_picker.dart';
+export 'src/fast_file_picker_path_extension.dart';