+++ lib/fc_file_picker_util.dart
@@ -1,2 +1,0 @@
-export 'src/fc_file_picker_util.dart';
-export 'src/fc_file_picker_path_extension.dart';