+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0
+
+- Return hasAccess in `useAppleScopedResource`.
+
 ## 0.0.1
 
 - Initial release